third_party/blink/web_tests/platform/generic/svg/as-object/embedded-svg-immediate-offsetWidth-query-expected.txt

Tests that querying offsetWidth twice leads to the same result - no red background should be visible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS newObject.offsetWidth is 0
PASS newObject.offsetWidth is 0
PASS newObject.offsetWidth is 302
PASS newObject.offsetWidth is 302
PASS newObject.offsetWidth is 402
PASS newObject.offsetWidth is 402
PASS successfullyParsed is true

TEST COMPLETE
